Get step parameters for SALSA

Get step parameters for SALSA

get_salsa_steps(x, include_vector = FALSE, ...)

Arguments

x
numeric vector of counts

...
additional arguments are ignored

Value

list containing "n_start", "step_size", "max_step", and when include_vector=TRUE it includes "n_vector" and "count_vector".

Details

This function takes a vector of counts and determines the appropriate step size to use when iterating the count threshold used by SALSA.

See also

Examples

library(salsa);
data(oz2_numi_per_cell);

usecounts <- sort(oz2_numi_per_cell$count);
get_salsa_steps(usecounts);

#> $n_start
#> [1] 1
#> 
#> $step_size
#> [1] 2
#> 
#> $max_step
#> [1] 96
#> 

# optionally return the vector of thresholds to use
get_salsa_steps(usecounts, include_vector=TRUE);

#> $n_start
#> [1] 1
#> 
#> $step_size
#> [1] 2
#> 
#> $max_step
#> [1] 96
#> 
#> $n_vector
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52 53 54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75
#> [76] 76 77 78 79 80 81 82 83 84 85 86 87 88 89 90 91 92 93 94 95 96
#> 
#> $count_vector
#>  [1]    0    3    8   15   24   35   48   63   80   99  120  143  168  195  224
#> [16]  255  288  323  360  399  440  483  528  575  624  675  728  783  840  899
#> [31]  960 1023 1088 1155 1224 1295 1368 1443 1520 1599 1680 1763 1848 1935 2024
#> [46] 2115 2208 2303 2400 2499 2600 2703 2808 2915 3024 3135 3248 3363 3480 3599
#> [61] 3720 3843 3968 4095 4224 4355 4488 4623 4760 4899 5040 5183 5328 5475 5624
#> [76] 5775 5928 6083 6240 6399 6560 6723 6888 7055 7224 7395 7568 7743 7920 8099
#> [91] 8280 8463 8648 8835 9024 9215
#>