Get SALSA results as a data.frame

Get SALSA results as a data.frame

get_salsa_table(fit_list, verbose = FALSE, ...)

Arguments

fit_list
list output from do_salsa_steps(), where each list element contains min_count, and one or more of fit_fr and fit_fr_wei.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Value

data.frame containing one row per minimum count threshold, with columns containing the Frechet and Frechet-Weibull fit parameters, or NA values when there are no parameters for the given minimum count threshold.

Details

This function converts the list of lists output from do_salsa_steps() into a "tidy" data.frame intended to be used for visualization.

See also

Examples

library(salsa);
data(oz2_numi_per_cell);
x <- oz2_numi_per_cell$count[oz2_numi_per_cell$count >= 16];
x_salsa <- do_salsa_steps(x,
   count_vector=c(16,32,128),
   cache_fr=NULL,
   cache_fr_wei=NULL);
x_df <- get_salsa_table(x_salsa);
x_df;

#>     count    shape     scale fr_weight fr_shape  fr_scale wei_shape wei_scale
#> 16     16 2.003168  25.18459      0.99      1.5  25.24369  1.514120  1775.742
#> 32     32 2.038619 138.20354      0.99      1.5 178.24735  1.500003  1700.841
#> 128   128 2.067224 265.69465      0.99      1.8 180.50922  1.500004  1699.818
#>     lower_bound upper_bound
#> 16     161.1956    102.8464
#> 32     480.9103    555.2840
#> 128    410.7675   1054.2652