Parameter bounds for Frechet-Weibull fit

Parameter bounds for Frechet-Weibull fit

params_fr_wei(param_fr_wei = NULL, ...)

Arguments

param_fr_wei
data.frame containing one or more colnames "fr_shape", "fr_scale", "wei_shape", "wei_scale", "fr_weight", and one or more rownames "start", "lower", "upper". Any non-NA values replace the corresponding default values.

...
additional arguments are ignored.

Value

matrix of five parameter upper, lower, and start values to be used by fitdist_fr_wei().

Details

This function defines a matrix with default parameter bounds to restrict the range of values allowed during the Frechet-Weibull distribution fit. Each of the five parameters are given start, lower, and upper values:

fr_weight the weight of Frechet relative to Weibull, on a scale of 0 to 1.

fr_shape the shape value for the Frechet distribution

fr_scale the scale value for the Frechet distribution

wei_shape the shape value for the Weibull distribution

wei_scale the scale value for the Weibull distribution

The output matrix can be modified then passed to do_salsa_steps() in order to customize the model fit.

See also

Examples

# default matrix of parameters
param_fr_wei <- params_fr_wei();
param_fr_wei;

#>       fr_weight fr_shape fr_scale wei_shape wei_scale
#> start      0.10      2.0      179       2.0      1700
#> lower      0.01      1.5        1       1.5         1
#> upper      0.99   1000.0    10000    1000.0     10000

# modify some values
custom_param_fr_wei_m <- rbind(
   start=c(fr_shape=50, wei_shape=40),
   lower=c(fr_shape=NA, wei_shape=400));
custom_param_fr_wei_m;

#>       fr_shape wei_shape
#> start       50        40
#> lower       NA       400

custom_param_fr_wei <- params_fr_wei(custom_param_fr_wei_m);
custom_param_fr_wei;

#>       fr_weight fr_shape fr_scale wei_shape wei_scale
#> start      0.10     50.0      179        40      1700
#> lower      0.01      1.5        1       400         1
#> upper      0.99   1000.0    10000      1000     10000