Get step parameters for SALSA

Get step parameters for SALSA

get_salsa_steps(x, include_vector = FALSE, ...)

Arguments

x
numeric vector of counts

...
additional arguments are ignored

Details

This function takes a vector of counts and determines the appropriate step size to use when iterating the count threshold used by SALSA.

Examples

library(salsa);
data(oz2_numi_per_cell);

usecounts <- sort(oz2_numi_per_cell$count);
get_salsa_steps(usecounts);

#> $n_start
#> [1] 1
#> 
#> $step_size
#> [1] 2
#> 
#> $n_bound
#> [1] 96
#> 

# optionally return the vector of thresholds to use
get_salsa_steps(usecounts, include_vector=TRUE);

#> $n_start
#> [1] 1
#> 
#> $step_size
#> [1] 2
#> 
#> $n_bound
#> [1] 96
#> 
#> $n_vector
#>  [1]  1  3  5  7  9 11 13 15 17 19 21 23 25 27 29 31 33 35 37 39 41 43 45 47 49
#> [26] 51 53 55 57 59 61 63 65 67 69 71 73 75 77 79 81 83 85 87 89 91 93 95
#> 
#> $count_vector
#>  [1]    0    8   24   48   80  120  168  224  288  360  440  528  624  728  840
#> [16]  960 1088 1224 1368 1520 1680 1848 2024 2208 2400 2600 2808 3024 3248 3480
#> [31] 3720 3968 4224 4488 4760 5040 5328 5624 5928 6240 6560 6888 7224 7568 7920
#> [46] 8280 8648 9024
#>